+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Minor*
 
 * Added `--include-seer` option to `start`. ([#179](https://github.com/hjwylde/werewolf/issues/179))
+* Added the Village Drunk role. ([#160](https://github.com/hjwylde/werewolf/issues/160))
 
 *Revisions*